Enhance Session Cookie handling
1. Modify Session Cookie security policy incase https is used.
2. Modify Session Cookie path to / with means it applies to all pages.

Change 2 is only a temporal modification for testing which needs to be revised in
the future. But this does what a naive user would expect which is set a cookie that
applies for the hole site not just for the subpages. however a user who is accustomend to the usage of
cookies might expect the other behaviour